API

Modules

BinnedModels.BinnedModels

Types and constants

BinnedModels.BinEdges

BinnedModels.BinnedModel

BinnedModels.Binning

Functions and macros

BinnedModels.bin_centers

BinnedModels.bin_intervals

BinnedModels.bin_leftedges

BinnedModels.bin_rightedges

BinnedModels.bin_widths

BinnedModels.binned_likelihood

BinnedModels.binned_model

Documentation

BinnedModels.BinnedModels — Module

BinnedModels

A Julia package for statistical modeling of binned data, e.g. histograms.

BinnedModels.BinEdges — Type

BinEdges{T<:Number} = AbstractVector{T}

Representation of a binning with n bins as n+1 edge points.

BinnedModels.BinnedModel — Type

struct BinnedModels.BinnedModel

Respresents a statistical model for binned data.

User code should not instantiate BinnedModel directly, but use binned_model instead.

BinnedModels.Binning — Type

Binning{T<:Number} = AbstractVector{<:AbstractInterval{T}}

Representation of a binning as a vector of intervals.

BinnedModels.bin_centers — Function

bin_centers(binnig::Binning)
bin_centers(edges::BinEdges)

Returns the bin centers of a binning defined by a vector of edges.

BinnedModels.bin_intervals — Function

function bin_intervals(edges::BinEdges, closed::Val = Val(:closedleft))::Binning

Returns a vector of n bin intervals, derived from a vector of n+1 bin edges.

BinnedModels.bin_leftedges — Function

bin_leftedges(edges::BinEdges)
bin_leftedges(binnig::Binning)

Returns the left edges of a binning defined by a vector of edges.

BinnedModels.bin_rightedges — Function

bin_rightedges(edges::BinEdges)
bin_rightedges(binnig::Binning)

Returns the right edges of a binning defined by a vector of edges.

BinnedModels.bin_widths — Function

bin_widths(binnig::Binning)
bin_widths(binning::BinEdges)

Returns the bin widths of a binning defined by a vector of edges.

BinnedModels.binned_likelihood — Function

binned_likelihood(f_expectation, edges::Tuple{AbstractVector,...}, data::AbstractVector{<:Integer})
binned_likelihood(f_expectation, h::StatsBase.Histogram{<:Integer})

Constructs a binned likelihood object that is compatible with the DensityInterface API.

See also binned_model.

BinnedModels.binned_model — Function

binned_model(f_expectation, edges::Tuple{AbstractVector,...})

Create a staticstical binned model from an density expectation function f_expectation and a tuple of bin edges edges.

See also binned_likelihood.