API

Modules

BinnedModels.BinnedModels

Types and constants

BinnedModels.BinEdges

BinnedModels.BinnedModel

BinnedModels.Binning

BinnedModels.BinningAlgorithm

BinnedModels.FixedNBins

BinnedModels.FreedmanDiaconisBinning

BinnedModels.RiceBinning

BinnedModels.ScottBinning

BinnedModels.SquareRootBinning

BinnedModels.SturgesBinning

Functions and macros

BinnedModels.bin_centers

BinnedModels.bin_intervals

BinnedModels.bin_leftedges

BinnedModels.bin_rightedges

BinnedModels.bin_widths

BinnedModels.binned_likelihood

BinnedModels.binned_model

BinnedModels.fit_binedges

Documentation

BinnedModels.BinnedModels — Module

BinnedModels

A Julia package for statistical modeling of binned data, e.g. histograms.

BinnedModels.BinEdges — Type

BinEdges{T<:Number} = AbstractVector{T}

Representation of a binning with n bins as n+1 edge points.

BinnedModels.BinnedModel — Type

struct BinnedModels.BinnedModel

Respresents a statistical model for binned data.

User code should not instantiate BinnedModel directly, but use binned_model instead.

BinnedModels.Binning — Type

Binning{T<:Number} = AbstractVector{<:AbstractInterval{T}}

Representation of a binning as a vector of intervals.

BinnedModels.BinningAlgorithm — Type

abstract type BinningAlgorithm

Abstract type for for algorithms that determine a suitable binning for a given set of data.

BinnedModels.FixedNBins — Type

FixedNBins(nbins::Int)

Selects a fixed number of bins.

Constructor: FixedNBins(; fields...)

Fields:

nbins::Int64

BinnedModels.FreedmanDiaconisBinning — Type

struct FreedmanDiaconisBinning <: BinningAlgorithm

Selects automatic binning based on the Freedman–Diaconis rule.

Constructor: FreedmanDiaconisBinning()

BinnedModels.RiceBinning — Type

struct RiceBinning <: BinningAlgorithm

Selects automatic binning based on the Rice rule.

Constructor: RiceBinning()

BinnedModels.ScottBinning — Type

struct ScottBinning <: BinningAlgorithm

Selects automatic binning based on Scott's normal reference rule.

Constructor: ScottBinning()

BinnedModels.SquareRootBinning — Type

struct SquareRootBinning <: BinningAlgorithm

Selects automatic binning based on the Square-root choice.

Constructor: SquareRootBinning()

BinnedModels.SturgesBinning — Type

struct SturgesBinning <: BinningAlgorithm

Selects automatic binning based on Sturges' formula.

Constructor: SturgesBinning()

BinnedModels.bin_centers — Function

bin_centers(binnig::Binning)
bin_centers(edges::BinEdges)

Returns the bin centers of a binning defined by a vector of edges.

BinnedModels.bin_intervals — Function

function bin_intervals(edges::BinEdges, closed::Val = Val(:closedleft))::Binning

Returns a vector of n bin intervals, derived from a vector of n+1 bin edges.

BinnedModels.bin_leftedges — Function

bin_leftedges(edges::BinEdges)
bin_leftedges(binnig::Binning)

Returns the left edges of a binning defined by a vector of edges.

BinnedModels.bin_rightedges — Function

bin_rightedges(edges::BinEdges)
bin_rightedges(binnig::Binning)

Returns the right edges of a binning defined by a vector of edges.

BinnedModels.bin_widths — Function

bin_widths(binnig::Binning)
bin_widths(binning::BinEdges)

Returns the bin widths of a binning defined by a vector of edges.

BinnedModels.binned_likelihood — Function

binned_likelihood(f_expectation, edges::Tuple{AbstractVector,...}, data::AbstractVector{<:Integer})
binned_likelihood(f_expectation, h::StatsBase.Histogram{<:Integer})

Constructs a binned likelihood object that is compatible with the DensityInterface API.

See also binned_model.

BinnedModels.binned_model — Function

binned_model(f_expectation, edges::Tuple{AbstractVector,...})

Create a staticstical binned model from an density expectation function f_expectation and a tuple of bin edges edges.

See also binned_likelihood.

BinnedModels.fit_binedges — Function

fit_binedges(data::AbstractVector{<:Real})::BinEdges
fit_binedges(data::NTuple{N,AbstractVector{<:Real}})::NTuple{N,BinEdges}

Return suitable bin edges for data, using algorithm.

data may be a real-valued vector, or a tuple of real-valued vectors for multi-dimensional data.